# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Task API",
    "version": "1.0.0",
    "description": "CRUD Task API with Supabase authentication (signup, login, logout, protected routes)."
  },
  "paths": {
    "/auth/signup": {
      "post": {
        "summary": "Create a new user",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "example": "test@example.com" },
                  "password": { "type": "string", "example": "password123" }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "User created" },
          "400": { "description": "Missing or invalid input" }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Log in and receive a JWT",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "example": "test@example.com" },
                  "password": { "type": "string", "example": "password123" }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Access token and refresh token" },
          "400": { "description": "Missing email or password" },
          "401": { "description": "Invalid login credentials" }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "summary": "Log out",
        "tags": ["Auth"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "204": { "description": "Logged out" },
          "401": { "description": "Missing or invalid token" }
        }
      }
    },
    "/protected/profile": {
      "get": {
        "summary": "Get the current user's profile",
        "tags": ["Protected"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "User profile (id, email, created_at)" },
          "401": { "description": "Missing or invalid token" }
        }
      }
    },
    "/protected/dashboard": {
      "get": {
        "summary": "Protected dashboard",
        "tags": ["Protected"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Dashboard message" },
          "401": { "description": "Missing or invalid token" }
        }
      }
    },
    "/public/info": {
      "get": {
        "summary": "Public info (no auth)",
        "tags": ["Public"],
        "responses": {
          "200": { "description": "Public message" }
        }
      }
    },
    "/tasks": {
      "get": {
        "summary": "Get all tasks",
        "tags": ["Tasks"],
        "responses": {
          "200": { "description": "A list of tasks" }
        }
      },
      "post": {
        "summary": "Create a new task",
        "tags": ["Tasks"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string", "example": "Buy milk" }
                },
                "required": ["title"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Task created" },
          "400": { "description": "Missing or empty title" }
        }
      }
    },
    "/tasks/{id}": {
      "get": {
        "summary": "Get a single task by id",
        "tags": ["Tasks"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": { "description": "The task" },
          "404": { "description": "Task not found" }
        }
      },
      "put": {
        "summary": "Update a task",
        "tags": ["Tasks"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string" },
                  "done": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Task updated" },
          "400": { "description": "Invalid body" },
          "404": { "description": "Task not found" }
        }
      },
      "delete": {
        "summary": "Delete a task",
        "tags": ["Tasks"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "204": { "description": "Task deleted" },
          "404": { "description": "Task not found" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}